requirements.txt
transformers>=4.31.0
datasets>=2.12.0
hydra-core>=1.3.2
omegaconf>=2.3.0
wandb>=0.15.4
tqdm>=4.65.0
accelerate>=0.20.3
bs4>=0.0.1
beautifulsoup4>=4.12.2
tensor-parallel>=1.3.0
sentencepiece>=0.1.99
protobuf>=4.23.3
fschat[model_worker,webui]
vllm
torchaudio
torchvision